backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
certifi==2025.10.5
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.1
h11==0.16.0
idna==3.10
motor==3.3.1
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
pymongo==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.15.0
uvicorn==0.25.0
